requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
anthropic
jieba
psutil
aiohttp
apscheduler
duckduckgo-search
aiofiles
python-multipart
baidu-aip
chardet
# v0.8.0 Phase 3: 长文本总结
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
pdf2image==1.17.0

# v0.9.0 Phase 1: 本地人脸识别 (预装依赖)
face_recognition
opencv-python-headless

# v0.9.0 Phase 2: 安全认证
python-jose[cryptography]
bcrypt

# Webhook 自动部署
flask
